roundtrip: Build ROM and compare to base ROM

+++ cli/tests/test_roundtrip.rs
@@ -8,7 +8,7 @@ use std::{
     process::Command,
 };
 
-use anyhow::Result;
+use anyhow::{Result, bail};
 use ds_decomp::{
     analysis::FindLocalDataError,
     config::{config::Config, module::ModuleError},
@@ -19,7 +19,7 @@ use ds_decomp_cli::{
 };
 use ds_rom::{
     crypto::blowfish::BlowfishKey,
-    rom::{Rom, raw},
+    rom::{Rom, RomLoadOptions, raw},
 };
 use log::LevelFilter;
 use zip::ZipArchive;
@@ -55,7 +55,10 @@ fn test_roundtrip() -> Result<()> {
         // Extract ROM
         let base_name = path.with_extension("").file_name().unwrap().to_str().unwrap().to_string();
         let project_path = roms_dir.join(&base_name);
-        let extract_path = extract_rom(&path, &project_path, &key)?;
+        let extract_path = project_path.join("extract");
+        let raw_rom = raw::Rom::from_file(path)?;
+        let rom = Rom::extract(&raw_rom)?;
+        rom.save(&extract_path, Some(&key))?;
         let rom_config = extract_path.join("config.yaml");
 
         // Init dsd project
@@ -74,7 +77,7 @@ fn test_roundtrip() -> Result<()> {
         })?;
         let dsd_config_yaml = dsd_config_dir.join("arm9/config.yaml");
         let dsd_config = Config::from_file(&dsd_config_yaml)?;
-        let target_config_dir = configs_dir.join(base_name);
+        let target_config_dir = configs_dir.join(&base_name);
         if allowed_unknown_function_calls {
             assert!(
                 target_config_dir.exists(),
@@ -147,6 +150,22 @@ fn test_roundtrip() -> Result<()> {
             ConfigRom { elf: linker_out_file.clone(), config: dsd_config_yaml.clone() };
         config_rom.run()?;
 
+        // Build ROM
+        let rom_config_path = dsd_config_dir
+            .join("arm9")
+            .join(&dsd_config.main_module.object)
+            .with_file_name("rom_config.yaml");
+        let rom_load_options = RomLoadOptions { key: Some(&key), ..Default::default() };
+        let rom = Rom::load(&rom_config_path, rom_load_options)?;
+        let built_rom = rom.build(Some(&key))?;
+        let rom_path = project_path.join(format!("build_{base_name}.nds"));
+        built_rom.save(rom_path)?;
+
+        // Compare ROMs
+        if built_rom.data() != raw_rom.data() {
+            bail!("Built ROM does not match base ROM");
+        }
+
         fs::remove_dir_all(project_path)?;
     }
 
@@ -191,14 +210,6 @@ fn dsd_init(
     };
     init.run()?;
     Ok(dsd_config_dir)
-}
-
-fn extract_rom(path: &Path, project_path: &Path, key: &BlowfishKey) -> Result<PathBuf> {
-    let extract_path = project_path.join("extract");
-    let raw_rom = raw::Rom::from_file(path)?;
-    let rom = Rom::extract(&raw_rom)?;
-    rom.save(&extract_path, Some(key))?;
-    Ok(extract_path)
 }
 
 fn directory_equals(target: &Path, base: &Path) -> Result<bool> {
